requirements.txt
# Core Dependencies
torch==2.1.2
torchvision==0.16.2
pytorch-lightning==2.1.3

# Model libraries
timm==0.9.8
efficientnet-pytorch==0.7.1

# Data processing
albumentations==1.4.3
scikit-image==0.22.0
opencv-python==4.8.1
pandas==2.1.3
numpy==1.26.2
Pillow==10.1.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1

# Explainability (XAI)
captum==0.7.0
shap==0.43.0
matplotlib==3.8.2
seaborn==0.13.1

# Deployment
streamlit==1.29.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Experiment tracking
wandb==0.16.3

# Hyperparameter tuning
optuna==3.14.0

# Utilities
scikit-learn==1.3.2
pyyaml==6.0.1
